# docs/modules/ROOT/attachments/openapi.yaml
openapi: 3.1.0
info:
  title: JSupervisor REST API
  description: |
    JSupervisor is a process management and monitoring service that provides
    comprehensive control over application processes. This API enables you to:

    - Start, stop, and restart processes individually or in bulk
    - Monitor process status and health
    - Retrieve process logs and statistics
    - Configure process behavior and auto-start settings

    The API follows RESTful principles and returns JSON responses.
  contact:
    name: Alex Mondshain
    url: https://www.alexmond.org/jsupervisor/current/
    email: alex.mondshain@gmail.com
  license:
    name: Apache License 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.txt
  version: 0.0.2
servers:
  - url: http://localhost:8086
    description: Local Development Server
...
paths:
...
  /api/v1/processes/group/stop:
    post:
      tags:
        - Group Process Operations
      summary: Stop all processes
      description: Stop all currently running processes
      operationId: stopAll
      responses:
        "200":
          description: All processes stopped successfully
        "500":
          description: Error occurred while stopping processes
        "404":
          description: No running processes found
  /api/v1/processes/group/start:
    post:
      tags:
        - Group Process Operations
      summary: Start all processes
      description: Start all configured processes that are not currently running
      operationId: startAll
      responses:
        "200":
          description: All processes started successfully
        "500":
          description: Error occurred while starting processes
        "409":
          description: Some processes are already running
  /api/v1/processes/group/restart:
    post:
      tags:
        - Group Process Operations
      summary: Restart all processes
      description: Restart all configured processes
      operationId: restartAll
      responses:
        "200":
          description: All processes restarted successfully
        "500":
          description: Error occurred while restarting processes
...
  /api/v1/events:
    get:
      tags:
        - Event Management
      summary: Get paginated events
      description: Retrieves a paginated list of all process events with customizable
        sorting
      operationId: getEvents
      parameters:
        - name: page
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            minimum: 0
          example: 0
        - name: size
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            maximum: 100
            minimum: 1
          example: 20
        - name: sortBy
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
            enum:
              - eventTime
              - processName
              - newStatus
              - oldStatus
              - pid
          example: eventTime
        - name: sortDirection
          in: query
          description: Sort direction
          required: false
          schema:
            type: string
            enum:
              - asc
              - desc
          example: desc
      responses:
        "200":
          description: Successfully retrieved events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page"
        "400":
          description: Invalid pagination or sorting parameters
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageProcessEventEntry"
  /api/v1/events/{id}:
    get:
      tags:
        - Event Management
      summary: Get event by ID
      description: Retrieves detailed information about a specific process event
      operationId: getEventById
      parameters:
        - name: id
          in: path
          description: Event ID
          required: true
          schema:
            type: integer
            format: int64
          example: 123
      responses:
        "200":
          description: Event found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProcessEventEntry"
        "404":
          description: Event not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProcessEventEntry"
...
  /api/v1/events/by-status/{status}:
    get:
      tags:
        - Event Management
      summary: Get events by status
      description: Retrieves paginated events matching a specific process status
      operationId: getEventsByStatus
      parameters:
        - name: status
          in: path
          description: Process status to filter by
          required: true
          schema:
            type: string
            enum:
              - not_started
              - running
              - finished
              - unknown
              - failed
              - failed_to_start
              - stopped
              - stopping
              - aborted
              - starting
              - healthy
              - unhealthy
          example: running
        - name: page
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            format: int32
            default: 0
          example: 0
        - name: size
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            format: int32
            default: 20
          example: 20
        - name: sortBy
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
            default: eventTime
          example: eventTime
        - name: sortDirection
          in: query
          description: Sort direction
          required: false
          schema:
            type: string
            default: desc
          example: desc
      responses:
        "200":
          description: Successfully retrieved events
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageProcessEventEntry"
        "400":
          description: Invalid status value
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageProcessEventEntry"
  /api/v1/events/by-process/{processName}:
    get:
      tags:
        - Event Management
      summary: Get events by process name
      description: Retrieves paginated events for a specific process
      operationId: getEventsByProcess
      parameters:
        - name: processName
          in: path
          description: Process name
          required: true
          schema:
            type: string
          example: myapp
        - name: page
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            format: int32
            default: 0
          example: 0
        - name: size
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            format: int32
            default: 20
          example: 20
        - name: sortBy
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
            default: eventTime
          example: eventTime
        - name: sortDirection
          in: query
          description: Sort direction
          required: false
          schema:
            type: string
            default: desc
          example: desc
      responses:
        "200":
          description: Successfully retrieved events
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageProcessEventEntry"
        "404":
          description: Process not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageProcessEventEntry"
components:
  schemas:
...
    ProcessEventEntry:
      type: object
      description: Represents an event entry for process state changes and lifecycle
        events
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the process event
          example: 1
        pid:
          type: integer
          format: int64
          description: Process ID of the running process
          example: 12345
        eventTime:
          type: string
          format: date-time
          description: Timestamp when the event occurred
          example: 2025-11-17T10:30:00
        processName:
          type: string
          description: Name of the managed process
          example: my-application
        startTime:
          type: string
          format: date-time
          description: Timestamp when the process was started
          example: 2025-11-17T10:00:00
        endTime:
          type: string
          format: date-time
          description: "Timestamp when the process ended, null if still running"
          example: 2025-11-17T10:15:00
        exitCode:
          type: integer
          format: int32
          description: "Exit code of the process, null if still running"
          example: 0
        newStatus:
          type: string
          description: New status of the process after the event
          enum:
            - not_started
            - running
            - finished
            - unknown
            - failed
            - failed_to_start
            - stopped
            - stopping
            - aborted
            - starting
            - healthy
            - unhealthy
        oldStatus:
          type: string
          description: Previous status of the process before the event
          enum:
            - not_started
            - running
            - finished
            - unknown
            - failed
            - failed_to_start
            - stopped
            - stopping
            - aborted
            - starting
            - healthy
            - unhealthy
        processUptime:
          type: string
          description: Duration for which the process has been running
          example: PT15M
...
    Page:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        first:
          type: boolean
        last:
          type: boolean
        size:
          type: integer
          format: int32
        content:
          type: array
          items: { }
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        pageable:
          $ref: "#/components/schemas/PageableObject"
        numberOfElements:
          type: integer
          format: int32
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: "#/components/schemas/SortObject"
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        unpaged:
          type: boolean
        paged:
          type: boolean
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
    PageProcessEventEntry:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        first:
          type: boolean
        last:
          type: boolean
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/ProcessEventEntry"
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        pageable:
          $ref: "#/components/schemas/PageableObject"
        numberOfElements:
          type: integer
          format: int32
        empty:
          type: boolean